Extendidas funciones de subida de imagenes y topografias a sistemas y a exploraciones
