+++ app/(instance-dashboard)/instance/[instance]/explorations/create/page.tsx
@@ -3,7 +3,7 @@ import BasicCard from '@/components/_Atoms/boxes/basic-card'
 import CardTitle from '@/components/_Atoms/boxes/card-title'
 import CollapsibleBox from '@/components/_Atoms/boxes/collapsible-box'
 import NotFoundCard from '@/components/cards/404-not-found'
-import ExplorationCreationForm from '@/components/_document-pages/exploration-cretion-form'
+import ExplorationCreationForm from '@/components/_document-pages/exploration-edition-board/exploration-cretion-form'
 import PageContainer from '@/components/theming/page-container'
 import {GroupIndex} from '@/database/services/group.actions'
 import {CaveIndex, getCaveIndex} from '@/database/services/cave.actions'
