+++ app/(instance-dashboard)/instance/[instance]/systems/[document]/edit/page.tsx
@@ -2,7 +2,7 @@ import {auth} from '@/auth'
 import BasicCard from '@/components/_Atoms/boxes/basic-card'
 import CardTitle from '@/components/_Atoms/boxes/card-title'
 import NotFoundCard from '@/components/cards/404-not-found'
-import SystemEditionForm from '@/components/_document-pages/system-edition-form'
+import SystemEditionForm from '@/components/_document-pages/system-edition-board/system-edition-form'
 import PageContainer from '@/components/theming/page-container'
 import {getPlainSystem, PlainSystem} from '@/database/services/system.actions'
 import {LuPlusCircle} from 'react-icons/lu'
