+++ app/(instance-dashboard)/instance/[instance]/systems/[document]/page.tsx
@@ -1,5 +1,4 @@
 import PageContainer from '@/components/theming/page-container'
-import ImageCard from '@/components/cards/image-card'
 import SystemInfoCard from '@/components/_document-pages/system-details-board/system-info-card'
 import SystemCavesCard from '@/components/_document-pages/system-details-board/system-caves-card'
 import SystemHeader from '@/components/_document-pages/system-details-board/system-header'
@@ -8,7 +7,8 @@ import SystemScienceCard from '@/components/_document-pages/system-details-board
 import SkeletonHeader from '@/components/cards/skelenton-header'
 import {Suspense} from 'react'
 import SkeletonCard from '@/components/cards/skeleton-card'
-import DocumentNotificationArea from '@/components/_document-pages/document-notification-area/document-notification-area'
+import PicturesLoader from '@/components/_document-pages/_shared-cards/pictures-loader'
+import TopographiesLoader from '@/components/_document-pages/_shared-cards/topographies-loader'
 
 interface PageProps {
   params: Promise<{document: string; instance: string}>
@@ -20,19 +20,17 @@ export default async function SystemDetailPage({params}: PageProps) {
 
   return (
     <PageContainer className="justify-start">
-      <Suspense fallback={null}>
-        <DocumentNotificationArea
-          instanceName={instance}
-          type="system"
-        />
-      </Suspense>
+      {/* Header */}
+      <div className="flex flex-col items-center w-full">
+        <Suspense fallback={<SkeletonHeader />}>
+          <SystemHeader
+            systemId={document}
+            instanceName={instance}
+          />
+        </Suspense>
+      </div>
 
-      <ImageCard />
-
-      <Suspense fallback={<SkeletonHeader />}>
-        <SystemHeader systemId={document} />
-      </Suspense>
-
+      {/* Content */}
       <div className="flex gap-4 flex-wrap justify-center">
         <Suspense fallback={<SkeletonCard />}>
           <SystemInfoCard systemId={document} />
@@ -48,6 +46,14 @@ export default async function SystemDetailPage({params}: PageProps) {
 
         <Suspense fallback={<SkeletonCard />}>
           <SystemCavesCard systemId={document} />
+        </Suspense>
+
+        <Suspense fallback={<SkeletonCard />}>
+          <PicturesLoader systemId={document} />
+        </Suspense>
+
+        <Suspense fallback={<SkeletonCard />}>
+          <TopographiesLoader systemId={document} />
         </Suspense>
       </div>
     </PageContainer>
