+++ app/(instance-dashboard)/instance/[instance]/systems/create/page.tsx
@@ -4,7 +4,7 @@ import BasicCard from '@/components/_Atoms/boxes/basic-card'
 import CardTitle from '@/components/_Atoms/boxes/card-title'
 import CollapsibleBox from '@/components/_Atoms/boxes/collapsible-box'
 import NotFoundCard from '@/components/cards/404-not-found'
-import SystemCreationForm from '@/components/_document-pages/system-creation-form'
+import SystemCreationForm from '@/components/_document-pages/system-edition-board/system-creation-form'
 import PageContainer from '@/components/theming/page-container'
 import {getInstanceId} from '@/database/services/instance.actions'
