+++ components/_document-pages/_shared-cards/pictures-card.tsx
