+++ components/_document-pages/_toolbars/remove-document-dialog.tsx
