--- components/_document-pages/cave-details-board/topographies-loader.tsx
@@ -1,9 +1,0 @@
-import {getPlainCave, PlainCave} from '@/database/services/cave.actions'
-import TopographiesCard from './topographies-card'
-
-export default async function TopographiesLoader({caveId}: {caveId: string}) {
-  // Obtener la cavidad
-  const cave = (await getPlainCave(caveId)).content as PlainCave | null
-
-  return <TopographiesCard topographies={cave?.topographies || []} />
-}
