+++ components/_document-pages/exploration-edition-board/exploration-cretion-form.tsx
@@ -4,27 +4,23 @@ import {useForm} from 'react-hook-form'
 import {zodResolver} from '@hookform/resolvers/zod'
 
 import {Answer} from '@/database/types/answer.type'
-import {
-  ExplorationFormValues,
-  explorationMaxCharacters,
-} from '@/database/validation/exploration.schemas'
+import {ExplorationFormValues} from '@/database/validation/exploration.schemas'
 import {ExplorationFormSchema} from '@/database/validation/exploration.schemas'
 import {GroupIndex} from '@/database/services/group.actions'
 
 import {Form} from '@/components/ui/form'
 import SubmitButton from '@/components/_Atoms/buttons/submit-button'
 import DbAwnserBox from '@/components/_Atoms/boxes/db-answer-box'
-import MultiRefSelectField from '@/components/_Atoms/fields/multi-ref-select-field'
-import TextField from '@/components/_Atoms/fields/text-field'
-import MultiTextField from '@/components/_Atoms/fields/multi-text-field'
-import TextAreaField from '@/components/_Atoms/fields/text-area-field'
-import MultiDateField from '@/components/_Atoms/fields/multi-date-field'
+import {Tabs, TabsContent, TabsList, TabsTrigger} from '@/components/ui/tabs'
+
 import LinkButton from '@/components/_Atoms/buttons/link-button'
 import {Button} from '@/components/ui/button'
-import TimeField from '@/components/_Atoms/fields/time-field'
 import {CaveIndex} from '@/database/services/cave.actions'
 import {createExploration} from '@/database/services/exploration.actions'
 import ReactHookFormErrorBox from '@/components/_Atoms/boxes/rhf-error-box'
+import {PiNumberCircleOneFill, PiNumberCircleTwoFill} from 'react-icons/pi'
+import ExplorationGeneralFormFragment from './1-general'
+import ExplorationPicturesFormFragment from './2-pictures'
 
 const EMPTY_EXPLORATION: ExplorationFormValues = {
   instances: [],
@@ -99,75 +95,37 @@ export default function ExplorationCreationForm({
         onSubmit={form.handleSubmit(onSubmit)}
         className="space-y-6"
       >
-        <MultiRefSelectField
-          control={form.control}
-          name="groups"
-          label="Grupos"
-          index={groupIndex}
-        />
-        <MultiRefSelectField
-          control={form.control}
-          name="caves"
-          label="Cavidades"
-          index={caveIndex}
-        />
-        <TextField
-          control={form.control}
-          name="name"
-          label="Titulo de la exploracion"
-          maxCharacters={explorationMaxCharacters.name}
-        />
-        <MultiDateField
-          control={form.control}
-          name="dates"
-          label="Fechas de la exploración"
-        />
+        <Tabs defaultValue="general">
+          <TabsList className="mx-auto h-auto mb-2 flex bg-transparent">
+            <TabsTrigger
+              value="general"
+              className="group data-[state=active]:bg-muted flex-1 flex-col p-3"
+            >
+              <PiNumberCircleOneFill className="group data-[state=active]:text-primary text-2xl rounded-full" />
+              General
+            </TabsTrigger>
 
-        <TimeField
-          control={form.control}
-          name="cave_time"
-          label="Tiempo en cueva"
-          description="Tiempo aproximado de trabajo dentro de cavidad"
-        />
+            <TabsTrigger
+              value="pictures"
+              className="group data-[state=active]:bg-muted flex-1 flex-col p-3"
+            >
+              <PiNumberCircleTwoFill className="text-2xl rounded-full" />
+              Imagenes
+            </TabsTrigger>
+          </TabsList>
+          <TabsContent value="general">
+            <ExplorationGeneralFormFragment
+              form={form}
+              caveIndex={caveIndex}
+              groupIndex={groupIndex}
+            />
+          </TabsContent>
 
-        <MultiTextField
-          control={form.control}
-          name="participants"
-          label="Participantes"
-          description="Introduce los participantes separados por enter"
-        />
-        <MultiTextField
-          control={form.control}
-          name="collaborators"
-          label="Colaboradores"
-          description="Introduce personas o entidades colaboradoras separadas por enter"
-        />
+          <TabsContent value="pictures">
+            <ExplorationPicturesFormFragment form={form} />
+          </TabsContent>
+        </Tabs>
 
-        <TextAreaField
-          control={form.control}
-          name="description"
-          label="Relato de la exploración"
-          maxCharacters={explorationMaxCharacters.description}
-        />
-
-        <TextAreaField
-          control={form.control}
-          name="incidents"
-          label="Incidentes ocurridos"
-          maxCharacters={explorationMaxCharacters.incidents}
-        />
-        <TextAreaField
-          control={form.control}
-          name="inventory"
-          label="Inventario de la exploración"
-          maxCharacters={explorationMaxCharacters.inventory}
-        />
-        <TextAreaField
-          control={form.control}
-          name="pending_work"
-          label="Trabajos pendientes"
-          maxCharacters={explorationMaxCharacters.pending_work}
-        />
         <ReactHookFormErrorBox errors={form.formState.errors} />
 
         <DbAwnserBox answer={dbAnswer} />
